# Conflicts:
#	CMakeLists.txt
#	include/roboteam_ai/control/positionControl/BBTrajectories/BBTrajectory1D.h
#	include/roboteam_ai/control/positionControl/BBTrajectories/BBTrajectory2D.h
#	include/roboteam_ai/control/positionControl/BBTrajectories/WorldObjects.h
#	include/roboteam_ai/control/positionControl/PositionControl.h
#	include/roboteam_ai/stp/plays/TestPlay.h
#	include/roboteam_ai/stp/tactics/TestTactic.h
#	src/ApplicationManager.cpp
#	src/control/ControlModule.cpp
#	src/control/positionControl/BBTrajectories/BBTrajectory1D.cpp
#	src/control/positionControl/BBTrajectories/BBTrajectory2D.cpp
#	src/control/positionControl/BBTrajectories/WorldObjects.cpp
#	src/control/positionControl/PositionControl.cpp
#	src/stp/plays/TestPlay.cpp
#	src/stp/skills/GoToPos.cpp
#	src/stp/skills/TestSkill.cpp
#	src/stp/tactics/TestTactic.cpp

+++ include/roboteam_ai/manual/JoystickHandler.h
@@ -2,12 +2,14 @@
 #ifndef RTT_JOYSTICKHANDLER_H
 #define RTT_JOYSTICKHANDLER_H
 
+#include <roboteam_proto/RobotCommand.pb.h>
+#include <stdio.h>
+
+#include <iostream>
+
+#include "JoystickState.h"
 #include "SDL.h"
 #include "SDL_joystick.h"
-#include <stdio.h>
-#include <iostream>
-#include "JoystickState.h"
-#include <roboteam_proto/RobotCommand.pb.h>
 
 namespace rtt::input {
 
@@ -17,7 +19,7 @@ class JoystickHandler {
     JoystickState joystickState;
     float robotAngle = 0.0;
     int robotId = -1;
-    int dribbler_vel = 0;
+    float dribbler_vel = 0;
     std::chrono::steady_clock::time_point id_switched_timestamp;
 
    public:
